+++ src/main.ts
@@ -2,7 +2,7 @@ import { NestFactory } from '@nestjs/core';
 import { NestExpressApplication } from '@nestjs/platform-express';
 import { AppModule } from './app.module';
 import { join } from 'path';
-import { HttpExceptionFilter } from './presentation/filter/httpExceptionFilter';
+import { HttpExceptionFilter } from './presentation/filters/httpExceptionFilter';
 
 /**
  * NestJSアプリケーションの起動
